Create licence.txt
Please excuse me for my rude licencing but this is the first time I have done something like this.
I just want to ensure that this project will keep mine. Also I want to ensure if somebody takes
this project whats totally fine he have to credit me and so on.
And as allways sorry for the bad english.